Update for N-API after landed in Node 8.x
This is a significant update to the napi_demo package and makes several
changes. Primary change is to update the leveldown package to a version
matching the N-API which landed in Node 8.x as experimental.

The app and website it starts are now much more dynamic. The website
updates the timestamp live such that subsequent entries can be added
to the database with different timestamps. Previously, we had to restart
Node.js to update the timestamp we had saved there.

Also moved around the app and site code to better isolate those
components. Now the site does all the data formatting while the app just
fetches raw values from the database.

App now supports serveral options (see '--help' for details) including
configuring the port and leveldb database path. This support relies on
the yargs package, which is now required.

Now includes a simple unit test which can be invoked via the --test
command line switch. This test just adds an entry and reads the database
back to make sure the entry was added. We output passed / failed
depending on results.